requirements.txt
comtypes==1.2.0
keyboard==0.13.5
Pillow==10.0.0
pynput==1.7.6
pystray==0.19.4
pywin32==306
pywinauto==0.6.8
six==1.16.0
